requirements.txt
apache-airflow-providers-amazon==9.0.0
pandas==2.2.3
